continue typing components
